requirements.txt
accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.1
blosc==1.11.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.0
decorator==5.1.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
eval_type_backport==0.2.0
exceptiongroup==1.2.1
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
gym==0.26.2
gym-minigrid==1.2.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.25.1
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipython==8.18.1
japanize-matplotlib==1.1.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai==1.34.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
protobuf==5.28.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyarrow-hotfix==0.6
pydantic==2.7.4
pydantic_core==2.18.4
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.14.0
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.66.4
traitlets==5.14.3
transformers==4.46.2
trl==0.8.6
typing_extensions==4.11.0
tyro==0.8.3
tzdata==2024.1
urllib3==2.2.1
wandb==0.18.1
wcwidth==0.2.13
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
